large patch

+++ main/correlation.py
@@ -9,25 +9,25 @@ def make_psd(W):
     return e_vec @ torch.diag(torch.clamp(e_val, min=1e-10)) @ e_vec.T
 
 
-def identity_cov(d, strength, device=config.DEVICE, dtype=torch.float64):
+def identity_cov(d, strength, device=config.DEVICE, dtype=torch.float32):
     return torch.eye(d, device=device, dtype=dtype)
 
 
-def exp_cov(d, strength, device=config.DEVICE, dtype=torch.float64):
+def exp_cov(d, strength, device=config.DEVICE, dtype=torch.float32):
     # task diversity K
     idx = torch.arange(d, device=device, dtype=dtype)
     W = strength ** torch.abs(idx[:, None] - idx[None, :])
     return make_psd(W)
 
 
-def powerlaw_cov(d, strength, device=config.DEVICE, dtype=torch.float64):
+def powerlaw_cov(d, strength, device=config.DEVICE, dtype=torch.float32):
     idx = torch.arange(d, device=device, dtype=dtype)
 
     W = 1.0 / (1.0 + torch.abs(idx[:, None] - idx[None, :])) ** strength
     return make_psd(W)
 
 
-def sample_task_weights(W, d, k, dtype=torch.float64):
+def sample_task_weights(W, d, k, dtype=torch.float32):
     e_val, e_vec = torch.linalg.eigh(W)
     W_sqrt = e_vec @ torch.diag(torch.sqrt(torch.clamp(e_val, min=0))) @ e_vec.T
     Z = torch.randn(k, d, device=W.device, dtype=dtype)
